Restructured: Added anthony gg tutorial.
